refactor(bench): re-home the four shared Node helpers out of Freehand
`lane_cache.cjs`, `navigate.cjs`, `sentinel.cjs` and `order_guard.cjs` are
pure Node (`node:fs`, `node:path`, `node:v8`) with zero Freehand coupling.
They were filed under `implementation/freehand/test/re_frame/freehand/bench/`
and would die with that tree.

They are NOT hicasso's. Four consumers outlive Freehand:

  implementation/scripts/compile-node-test.cjs        the script that
                                                      compiles every
                                                      node-test bundle
  implementation/core/test/re_frame/bench/p0_run.cjs
  adapters/reagent/.../bench/hicasso_narrow_run.cjs
  implementation/hicasso/test/re_frame/bench/hicasso/**

Homing them in `implementation/hicasso/` would point core's bench, Reagent's
bench and compile-node-test.cjs into a package heading for publication.

They go to `implementation/core/test/re_frame/bench/`, which already IS the
substrate-neutral bench home — `p0_run.cjs` is a Node driver of exactly this
shape, and `re-frame.bench.order-guard` is the CLJS/JVM half of the very rule
`order_guard.cjs` carries. No new home is added. The `.cljc` docstring's
reason for the split ("core may not `:require` out of `implementation/freehand`")
is dissolved by the move and is replaced with the durable one: the measured
page cannot load a driver-side CommonJS module.

All four files move verbatim; only path strings in callers change.

`lane_cache_wiring.test.cjs` pinned `require('./lane_cache.cjs')` with a
literal `./`. The helper is now reached at five different relative depths, so
the prefix is unpinned; the check still owns THAT a rider imports the cache
rule. Its four fixtures carry the real rider path so they keep modelling a
shape a driver can actually take.

Refs rf2-it4y5.
