+++ implementation/freehand/test/re_frame/freehand/bench/lane_cache_fixtures/agreeing_ids_run.cjs
@@ -5,7 +5,7 @@
 // it. Read as TEXT and never executed — see the fixtures note in the gate.
 const path = require('node:path');
 const { spawnSync } = require('node:child_process');
-const { resetLaneBuildCache } = require('./lane_cache.cjs');
+const { resetLaneBuildCache } = require('../../../../../core/test/re_frame/bench/lane_cache.cjs');
 
 const IMPL = path.resolve(__dirname, '..', '..', '..', '..', '..');
 const BUILD = 'freehand-release';
