+++ implementation/freehand/test/re_frame/freehand/bench/lane_cache_fixtures/string_clear_run.cjs
@@ -14,7 +14,7 @@
 // this file on the half it missed.
 const path = require('node:path');
 const { spawnSync } = require('node:child_process');
-require('./lane_cache.cjs');
+require('../../../../../core/test/re_frame/bench/lane_cache.cjs');
 
 const IMPL = path.resolve(__dirname, '..', '..', '..', '..', '..');
 const BUILD = 'freehand-release';
