+++ implementation/hicasso/test/re_frame/bench/hicasso/compile_gate.cjs
@@ -158,7 +158,7 @@ const fs = require('node:fs');
 const path = require('node:path');
 
 const { shadowBuild } = require('./lane_build.cjs');
-const { resetLaneBuildCache } = require('../../../../../freehand/test/re_frame/freehand/bench/lane_cache.cjs');
+const { resetLaneBuildCache } = require('../../../../../core/test/re_frame/bench/lane_cache.cjs');
 
 const LANE_DIR = __dirname;
 const IMPL = path.resolve(__dirname, '../../../../..');
