+++ implementation/hicasso/test/re_frame/bench/hicasso/hd8_clock_run.cjs
@@ -99,12 +99,12 @@ const fs = require('node:fs');
 const http = require('node:http');
 const path = require('node:path');
 
-const { navigate, NAV_TIMEOUT_MS } = require('../../../../../freehand/test/re_frame/freehand/bench/navigate.cjs');
-const { resetLaneBuildCache } = require('../../../../../freehand/test/re_frame/freehand/bench/lane_cache.cjs');
+const { navigate, NAV_TIMEOUT_MS } = require('../../../../../core/test/re_frame/bench/navigate.cjs');
+const { resetLaneBuildCache } = require('../../../../../core/test/re_frame/bench/lane_cache.cjs');
 // shadow-cljs exits 0 on WARNINGS, so a status check is not a gate. The
 // lane's one build door refuses a warned build (rf2-2rtt6.73).
 const { shadowBuild } = require('./lane_build.cjs');
-const guard = require('../../../../../freehand/test/re_frame/freehand/bench/order_guard.cjs');
+const guard = require('../../../../../core/test/re_frame/bench/order_guard.cjs');
 const seamlib = require('./seam.cjs');
 
 const IMPL = path.resolve(__dirname, '../../../../..');
